Handling SSL Termination
In a scenario where the Application Server (AS) is being a reverse
proxy it is very common that it's the reverse proxy that handles
the SSL Termination and then creates a HTTP connection to the AS.

The WSDL generation for providing WSDL for available endpoints at
the AS looks into the request's server port and server scheme sadly
it does not look by itself into x-forward-proto and x-forward-port
proxy headers.

This commit provides a filter that wraps the request to be aware of
SSL Termination at the reverse proxy, this is only mapped into the
/web-services/* URL mapping since for now we only have this need in
the web-services. But being this a generic filter it may, in the
future, be promoted into a product filter.

This filter still needs that the reverse proxy adds the
x-forward-port and x-forward-proto headers to the request before
forwarding it to the AS.

How can you do that?

In Haproxy in your backend configuration add:

http-request set-header x-forwarded-proto https
http-request set-header x-forwarded-port  443

In Apache server in your proxy configuration add:

RequestHeader set x-forwarded-proto https
RequestHeader set x-forwarded-port "443"

In Nginx when configuring your location for proxy pass add:

proxy_set_header x-forward-proto https;
proxy_set_header x-forward-port 443;

Relates with: #qubIT-Fenix-4235